Relocates Schema Providing and Descriptor obtaining methods to BigQuerySchemaProvider class.
